+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### 2.59
 - Private events are now reminders and are available in regional burns
+- Mutant Vehicles show location as "Out on the Playa"
 
 ### 2.58